+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Accessing the media type of a file no longer raise an error when it's unavailable, instead it defaults to `application/octet-stream`
 - When adding a `Directory` loaded via `Adapter\Filesystem` the sub directories are no longer loaded if not modified
+- Throw an exception when failing to load a file
 
 ## 6.3.2 - 2023-04-30